+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 ### Features
 
+- Добавлена middleware `cookie-parser`, для удобной работы с куками (инициализируется в методе `RestApplication.initCookieParser`)
+- Добавлено поле `IRestAppModuleConfig.cookieSecret` для возможности подписи кук
 - Переход с библиотеки `@ntegral/nestjs-sentry` на `@sentry/nestjs`
 - Добавлена проверка сложности пароля в `PasswordField` ([#223](https://gitlab.kozhindev.com/steroids/steroids-nest/-/issues/223))
 - Добавлен параметр `isListenLocalhost` в `RestApplication` для прослушивания только localhost'а, env-переменная для него - `APP_LISTEN_LOCALHOST` ([#215](https://gitlab.kozhindev.com/steroids/steroids-nest/-/issues/215))
-- Добавлена middleware `cookie-parser`, для удобной работы с куками
-- Добавлено поле `IRestAppModuleConfig.cookieSecret` для возможности подписи кук
 
 ### Fixes